LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-buttons-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>IOSInlineMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new IOSInlineMediaControls({ width: 680, height: 300 });
document.body.appendChild(mediaControls.element);

const buttonPositions = [
    [mediaControls.playPauseButton, 12],
    [mediaControls.skipBackButton, 10],
    [mediaControls.airplayButton, 13],
    [mediaControls.pipButton, 13],
    [mediaControls.fullscreenButton, 13]
];

let style;
shouldBecomeEqual("buttonPositions.some(buttonPosition => buttonPosition[0].element.style.width == '')", "false", () => {
    requestAnimationFrame(() => {
        for (let [button, y] of buttonPositions) {
            debug("");
            style = window.getComputedStyle(button.element);
            debug(button.constructor.name);
            shouldBeEqualToRGBAColor("style.backgroundColor", "rgb(0, 0, 0)");
            shouldBeEqualToString("style.height", "50px");
            shouldBeEqualToString("style.webkitMaskPositionY", `${y}px`);
        }

        debug("");
        mediaControls.element.remove();
        finishMediaControlsTest();
    });
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
