LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-buttons-styles-expected.txt

Testing the IOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS buttonPositions.some(buttonPosition => buttonPosition[0].element.style.width == '') became false

PlayPauseButton
PASS rgba(style.backgroundColor).r is 0
PASS rgba(style.backgroundColor).g is 0
PASS rgba(style.backgroundColor).b is 0
PASS rgba(style.backgroundColor).a is within 0.001 of 0
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "12px"

SkipBackButton
PASS rgba(style.backgroundColor).r is 0
PASS rgba(style.backgroundColor).g is 0
PASS rgba(style.backgroundColor).b is 0
PASS rgba(style.backgroundColor).a is within 0.001 of 0
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "10px"

AirplayButton
PASS rgba(style.backgroundColor).r is 0
PASS rgba(style.backgroundColor).g is 0
PASS rgba(style.backgroundColor).b is 0
PASS rgba(style.backgroundColor).a is within 0.001 of 0
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

PiPButton
PASS rgba(style.backgroundColor).r is 0
PASS rgba(style.backgroundColor).g is 0
PASS rgba(style.backgroundColor).b is 0
PASS rgba(style.backgroundColor).a is within 0.001 of 0
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

FullscreenButton
PASS rgba(style.backgroundColor).r is 0
PASS rgba(style.backgroundColor).g is 0
PASS rgba(style.backgroundColor).b is 0
PASS rgba(style.backgroundColor).a is within 0.001 of 0
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

PASS successfullyParsed is true

TEST COMPLETE
